+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `Versions`:
     * `Klock`: `2.0.0` -> `2.0.1`
+* `Repo`
+    * Repo `WriteStandardKeyValueRepo` got new method `unsetWithValues`
 
 ## 0.4.9